Fix Issues stream
The Jira API made a few changes on how the Issues stream is fetched. This led to forever running jobs that don't manage to load the data into Snowflake. Changing the pagination and the replication key to timestamp field instead of id fixed the bad behaviour and even improved the performance.